Rearrange shalwater_depth assignment to inside WRF
An issue was found when shalwater_depth was used to overwrite
WATER_DEPTH within real.exe. If shalwater_depth was assigned
at this point, then it could be overwritten when running WRF
and that may be confusing to follow / give users unexpected
results. Thus, making is so WATER_DEPTH = 0.0 if there is no
bathymetry data and then assigning the shalwater_depth value
in WRF makes it so that it only exists in one part of the code