Progect MYNN PBL from being used with LES PBL
TYPE: bug fix

KEYWORDS: PBL, MYNN, LES

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
The MYNN PBL schemes are not to be used on a coarse grid when the LES scheme is used on
a finer grid.

Solution:
A controlled stop is put in place when these namelist options are selected.

LIST OF MODIFIED FILES:
modified:   Registry/Registry.EM_COMMON
modified:   share/module_check_a_mundo.F

TESTS CONDUCTED:
1. When the MYNN and LES are selected, the code successfully stops.
2. Jenkins is all PASS.

RELEASE NOTE: A test was introduced to stop the combination of either of the MYNN PBL schemes on coarser grids with the LES PBL option selected on a finer grid.