submodule update in physics Makefile
In the physics Makefile, before any files source files are processed,
check to see if the NoahMP source files exist (the *.F files). If so,
then proceed as normally (they must already be linked in). If the files
do not exist, then `git submodule update --init --recursive` to bring in
the NoahMP repository, and then manually link the files into the
original WRF locations. If the source files did not exist, assume that the
NoahMP table files were also missing, so put those in the run directory.

Since these files are not really part of the WRF repo anymore, we remove
them when asked for a super-duper cleaning: `clean -aa`.

M clean
M phys/Makefile