Locate NoahMP src and tables into WRF directories
In the physics Makefile, before any files source files are processed,
check to see if the NoahMP source files exist (the *.F files). If so,
then proceed as normally (they must already be linked in). If the files
do not exist, then `git submodule update --init --recursive` to bring in
the NoahMP repository, and then manually link the files into the
original WRF locations. If the source files did not exist, assume that the
NoahMP table file (MPTABLE.TBL) is also missing, so put that file in the
run directory.

Since these files are not really part of the WRF repo anymore, we remove
them when asked for a super-duper cleaning: `clean -aa`.

An "if" test in the phys/Makefile detects if the NoahMP source files
exist. If not, then the git submodule command is used. The important
item is that the top-level Makefile has a submodules target that is processed
for each build.

M Makefile
M clean
M phys/Makefile