Fix un-initialized variable BHTMP in NTU microphysics (#1678)
TYPE: bug fix

KEYWORDS: BHTMP, mp_ntu

SOURCE: Tzu-Chin Tsai (National Taiwan University), reported by CWB

DESCRIPTION OF CHANGES: The real variable of BHTMP, part of the 2nd term for hail ventilation, was not assigned in the small_dt module (for vapor diffusion process) of ntu scheme.

Solution:
The value of BHTMP is defined in line 3831.

LIST OF MODIFIED FILES:
M phys/module_mp_ntu.F

TESTS CONDUCTED:
One idealized 2D squall-line test was performed, and the mean hail diameter was slightly increased from 1.348 (with the bug) to 1.369 (fixed) mm by 1.56% (the figure not shown). So, the impact of this bug on the simulation could be minor.