Update VEGPARM.TBL for NLCD40
WRFv4.4 crashes upon LSM initialization when NLCD40 landuse dataset is used. This crash was found to be related to the LCZ_* specs at the end of the main NLCD40 table entries. We just copied the values used for MODIS since the NLCD40 is a blend of US NLCD and MODIS.