Update da_read_obs_bufrseviri.inc
Meteosat 11 was added to the GFS bufr files in Feb 2018 meaning when the seviri bufr read was turned on in the namelist it would throw out an error message for each data point in the log file making it very large. The change to the error loop removes this message but also prepares the bufr read file to ingest meteosat 11 data into wrf. Dont think a crtm currently supports meteosat 11 but i can change the VARBC.in file to include data for this sat number and add a msg-4-seviri.info file to the radiance_info folder in another commit if this might be desirable for people to use.