Update namelist README for VAR to include use_clddet change (#1778)
TYPE: readme

DESCRIPTION OF CHANGES:
Problem:
Since V4.2 the cloud detection setting has changed with a change of registry values to decommission individual set mmr and ecmwf to include a single value that chooses one of 3 schemes to use. Also addition of a cloud detection scheme from Zhunge and Zou which has an individual place holder in the namelist.

Solution:
Changed readme file in VAR to reflect the change that happened in V4.2.

LIST OF MODIFIED FILES: var/README.namelist

TESTS CONDUCTED:
No tests conducted with no change to model.