Adjust defines and registry check to only look for defs not defs with value
This now mirrors the make build and allows the make build to supply
the WRF_CHEM and WRF_KPP option via defines/command line similar to
the make build. Likewise, while there are files that use
or
or any variation of this, these were never being used by the make build.
This may be a bug, but the priority of the cmake build is to acheive
parity with the make build in compile flags and overall effect.