Subroutine for PnetCDF non-blocking buffersize calc
This commits add a subroutine "BputCalcBufferSize". It calculates the
buffer size needed by PnetCDF non-blocking APIs (bput calls). The
returned buffer size will equal the total amount of writes (in bytes)
for a particular timestep, excluding PnetCDF headers (only the data
section).

PnetCDF bput calls need additional buffer space because they cache write
reqs in memory and flush them to file system altogether later on. The
additional buffer space is for the caching purpose.

----------

Below explains the implementaion of this new subroutine.

The implementation codes are borrowed from
share/output_wrf.F: subroutine "output_wrf".

The subroutine "output_wrf" is to write all variables to file.
The implementation logic of "output_wrf" is that:

  Iterate over all variables (do while loop):
      Some "if" conditions:
          write/output the variable (of size amoutX) to file
      End IF
  End loop.

The new subroutine "BputCalcBufferSize" is to return the buffer size
(i.e. the total amount of writes). The logic here is to replace
"write/output the variable to file" in the original logic with
"increment total size":

  totalSize = 0
  Iterate over all variables (do while loop):
      Some "if" conditions:
          totalSize = totalSize + amoutX
      End IF
  End loop.