PnetCDF: enter define mode only once
This is an optimization that applies to both blocking and non-blocking
PnetCDF APIs.

In PnetCDF, an application creating a file will first enter
"define mode", in which it can describe all attributes, dimensions,
types and structures of variables. The program will then exit
"define mode" and enter data mode, in which it actually performs I/O.

Previously in WRF, it enters and exits define mode several times,
while only once is enough/necessary. This commit remove redundant enter/
exit of define mode.