PnetCDF: only rank 0 write non-partitioned var
This is an optmization for PnetCDF (both blocking and non-blocking APIs)
. In WRF, there are two types of variables, paritioned and
non-partitioned. Non-partitioned variables are small variables that all
process write the entire variables with the same values to the file.
Partitioned varaibles are large variables; each process is responsible
for a (non-overrlapped) sub-region of such a variable.

When writing a non-partitioned variable, there's no need for every
process to make the PnetCDF write call.
1) If non-blocking APIs are used, only rank 0 will post write reqs
   using nfmpi_bput. Other processes do not need to call nfmpi_bput
   because nfmpi_bput does not require collective call.
2) If using blocking APIs, all processes still need to call
   nfmpi_put_all (which requires collective call). But only process of
   rank 0 has req size > 0, all other processes have req size = 0 (i.e.
   vcount = 0)