Add a simple test definition for testing GNU make builds
Following the documentation of the hpc-workflows testing framework and the
testing structure found in .ci/, a JSON file for a GNU compilation test was added.
This test will compile the em_real core using the GNU Linux x86 stanza configuration.

All other options are left as default. If this test is run using the derecho
configuration the appropriate modules will attempt to be loaded. For non-derecho
environments, per the testing structure under .ci/, if no configuration exists in
.ci/hostenv.sh then the current environment wil be used verbatim.