a few changes in documenting namelists (#2074)
TYPE: text only

KEYWORDS: namelists, corrections

SOURCE: JiriRichter and internal

DESCRIPTION OF CHANGES:
The description of a few namelists are updated. io_style_emiss is removed from this file since it is for WRF-Chem, and not specifiable under &time_control. Namelist ntiedtke_dx_opt was also removed, since it is not implemented.

ISSUE: For use when this PR closes an issue.
Fixes #2071, #2072, #2073

LIST OF MODIFIED FILES:
run/README.namelist

RELEASE NOTE: The documentation of a few namelists is updated in this PR.