Breakout deallocation calls into simpler smaller files (#2070)
TYPE: enhancement

KEYWORDS: intel, compilation, llvm, memory

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
The Intel oneAPI compilers (and others like nvhpc) struggle with some of
the larger (15k+ lines of code) files within WRF. This causes intense
memory usage that is not often available to the average user not in a
resource-rich environment. This often limits compilation to single
threaded if even possible or to a dedicated environment with enough
memory if available. If neither of those is available to a user, they
will be unable to use these configurations entirely.

Solution:
This PR focuses on the `deallocs.inc` sections of code used in
`module_domain` to reduce the include size to manageable levels. The
include is instead broken out into many smaller files as external
subroutines. The files are fully generated source code from the
registry, with the calls to the subroutines also being generated as
well. This also makes it relatively easy to change the number of files
generated from a source code perspective. Build rules would need to be
modified accordingly as seen in these changes.

TESTS CONDUCTED: 
Attached to this PR are plots of the respective effects of theses
changes. Changes were tested with intel and gcc compilers, but only
intel memory usage is shown as it exacerbates the memory usage issue.