CMake improve configure_reader.py script (#2087)
TYPE: enhancement

KEYWORDS: cmake, configuration

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
In preparation for alternate core compilation, there is need for
additional information from the stanzas. Some of these fields include
extraneous information to cmake that must be filtered out. The previous
version manually listed fields and values to sanitize, but this is
brittle when accounting for all stanza types and configurations.

Solution:
Rather than hard-coding the sanitization of stanza fields to be consumed
by cmake, a generalized approach is used to allow full use of stanza
fields down the line.