Use commit statuses to reinform PRs of run state
The input trigger or default push shall inform which
SHA to use, which will effectively reinform the PR of
the state of the workflow run. This now allows the
workflow to deliberately override or post workflow
runs to a PR without implicit events removing previous
statuses or skipped jobs polluting the PR state.