Find module works for configuration but can result in dangerous linking
When PnetCDF is configured to use netCDF, the find module will
appropriately locate this netCDF if possible to identify via
pnetcdf-config information, but the current built then continues
to find netCDF-Fortran. This may be co-located with a different
version of netCDF which may result in conflicting linking of
the netCDF library where used.