Revert cmake min version back to 3.20
Other parts of the cmake build rely on language
overrides being fed into the compiler. While this
could be adjusted to maintain the lower version
number, these adjustments to the source code are
outside the scope of these changes. Coupled with
the only-recent version drop, the more appropriate
approach is to revert the version back to the
higher working version and evaluate the necessary
code adjustments for cmake 3.19 to be viable at
a later time in a separate feature.