Restructure search for appropriate libdir substitute for nf-config
As nf-config does not provide an equivalent --libdir output we must
derive the info ourselves. --flibs provided the info previously, but it
may be more practical to use the --prefix output and search suspected
suffixes instead as --flibs contains other libraries' link info. To
handle the cases where --prefix is empty, we still fall back to --flibs
and if that proves insufficient then the as a last resort one directory
above the include directory can be used as a substitute for the prefix.