Fix errant line contiuation syntax in chem code (#2223)
TYPE: bug fix

KEYWORDS: syntax

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
There is invalid Fortran syntax for line continuation in the chem code.
More sensitive compilers catch this as an error during the CMake build
as files are not sanitized as much compared to the make build.

Solution:
Change the lines to be a single line. 

ISSUE: 
#2214 

TESTS CONDUCTED: 
1. Recreated issue noted in #2214 using Intel OneAPI compilers and
testing code modifications to pass compilation successfully.